Se hacen cambios a reservas.js y se crea index.js

+++ js/reservas.js
@@ -9,7 +9,7 @@ function getCottageInfo () {
     }else{
         return [];
     }
-}
+};
 
 function getExpInfo() {
     const experienceBookingLS = localStorage.getItem("experienceBookingInformation");
@@ -19,7 +19,7 @@ function getExpInfo() {
     }
 
     return [];
-}
+};
 
 //VARIABLES GLOBALES - GLOBAL VARIABLES
 
@@ -45,20 +45,68 @@ const inputCheckIn = document.getElementById("check-in");
 const inputCheckOut = document.getElementById("check-out");
 const inputExpDate = document.getElementById("exp-date");
 
+//SE CONFIGURA FECHA PARA QUE LA RESERVA NO PUEDA SER ANTERIOR AL DÍA PRESENTE 
+
+inputCheckIn.min = new Date().toISOString().split("T")[0];
+inputCheckOut.min = new Date().toISOString().split("T")[0];
+inputExpDate.min = new Date().toISOString().split("T")[0];
+
 
 //EVENTO DEL BOTÓN CABAÑA - COTTAGE BUTTON EVENT
 
 cottageBtn.addEventListener("click", (_Event) => {
+    //SEGÚN EL BOTÓN EN QUE SE HAGA CLICK, APARECEN LAS OPCIONES DE RESERVA
     cottageBookingForm.classList.add("activo");
     experienceBoookingForm.classList.remove("activo");
 
+    //SELECCIÓN DE CABAÑAS
+
+    selectCottage.addEventListener ("change", (_Event) => {
+        switch(selectCottage.value){
+            case "Cabaña Escondida":
+                cottageEscondida.reservation();
+
+                inputCheckOut.addEventListener("change", (_event) => {
+                    cottageEscondida.fullPrice();
+                });
+            break;
+
+            case "Cabaña Fagnano":
+                cottageFagnano.reservation();
+
+                inputCheckOut.addEventListener("change", (_event) => {
+                    cottageFagnano.fullPrice();
+                });
+            break;
+
+            case "Cabaña Roca":
+                cottageRoca.reservation();
+
+                inputCheckOut.addEventListener("change", (_event) => {
+                    cottageRoca.fullPrice();
+                });
+            break;
+
+            case "Cabaña Esmeralda":
+                cottageEsmeralda.reservation();
+
+                inputCheckOut.addEventListener("change", (_event) => {
+                    cottageEsmeralda.fullPrice();
+                });
+            break;
+        }
+    });
+
+
     const cottageBookingInformation = getCottageInfo();
+
+    //EVENTO PARA ENVIAR INFORMACIÓN DE RESERVA
 
     cottageBookingForm.addEventListener("submit", (Event) => {
 
         //STOP EVENT
         Event.preventDefault();
-        
+    
         //SE OBTIENEN DATOS DEL FORMULARIO - GETTING INPUT DATA
         const cottage = selectCottage.value;
         const guestName = inputGuestName.value;
@@ -69,9 +117,15 @@ cottageBtn.addEventListener("click", (_Event) => {
         const checkIn = inputCheckIn.value;
         const checkOut = inputCheckOut.value;
     
-    
-        //SE HACE PUSH DE LOS DATOS DE LA RESERVA AL ARRAY - PUSH INPUT DATA INTO AN ARRAY
-        cottageBookingInformation.push({
+        if (cottage == "" || guestName == "" || guestTelephone == "" || guestEmail == "" || checkIn == "" || checkOut == ""){
+
+            //MENSAJE DE FALTA DE DATOS
+            Swal.fire('Por favor, complete todos los campos.')
+
+        }else{
+
+            //SE HACE PUSH DE LOS DATOS DE LA RESERVA AL ARRAY - PUSH INPUT DATA INTO AN ARRAY
+            cottageBookingInformation.push({
             cabaña: cottage,
             nombre: guestName,
             email: guestEmail,
@@ -80,46 +134,97 @@ cottageBtn.addEventListener("click", (_Event) => {
             cantidadDeMenores: children,
             ingreso: checkIn,
             salida: checkOut,
-        });
+            });
 
-        //SE CARGA EL ARRAY AL LOCALSTORAGE - ARRAY IS UPLOADED TO LOCALSTORAGE
-        localStorage.setItem("cottageBookingInformation", JSON.stringify(cottageBookingInformation));
+            //SE CARGA EL ARRAY AL LOCALSTORAGE - ARRAY IS UPLOADED TO LOCALSTORAGE
+            localStorage.setItem("cottageBookingInformation", JSON.stringify(cottageBookingInformation));
 
+            //MENSAJE RESERVA COMPLETA
+            Swal.fire({
+                position: 'center',
+                title: 'Su reserva se ha realizado.',
+                showConfirmButton: true,
+            });
+        
+            //SE LIMPIA EL COMUNICADOR DE PRECIOS
+            let container = document.getElementById("container-precios");
+            container.innerHTML = "";
 
-        alert("Su reserva ha sido realizada.")
+            //SE LIMPIAN LOS INPUTS
+            selectCottage.value = "";
+            inputGuestName.value = "";
+            inputGuestEmail.value = "";
+            inputGuestTelephone.value = "";
+            inputAdults.value = "";
+            inputChildren.value = "";
+            inputCheckIn.value = "";
+            inputCheckOut.value = "";
 
-
-        //SE LIMPIAN LOS INPUTS
-        selectCottage.value = "";
-        inputGuestName.value = "";
-        inputGuestEmail.value = "";
-        inputGuestTelephone.value = "";
-        inputAdults.value = "";
-        inputChildren.value = "";
-        inputCheckIn.value = "";
-        inputCheckOut.value = "";
-
-        console.log(cottageBookingInformation);
-});
-
+            console.log(cottageBookingInformation);
+        };
+    });
 });
 
 
 //EVENTO DEL BOTÓN EXPERIENCIA - EXPERIENCE BUTTON EVENT
 
 experienceBtn.addEventListener("click", (_Event) => {
+    //SEGÚN EL BOTÓN EN QUE SE HAGA CLICK, APARECEN LAS OPCIONES DE RESERVA
     cottageBookingForm.classList.remove("activo");
     experienceBoookingForm.classList.add("activo");
 
+    //SELECCIÓN DE CABAÑAS
+
+    selectExperience.addEventListener ("change", (_Event) => {
+        switch(selectExperience.value){
+            case "Laguna Esmeralda":
+                experienceLagunaEsmeralda.reservation();
+
+                inputExpAdults.addEventListener("change", (_event) => {
+                    experienceLagunaEsmeralda.fullPrice();
+                });
+                
+                inputExpChildren.addEventListener("change", (_event) => {
+                    experienceLagunaEsmeralda.fullPrice();
+                });
+            break;
+
+            case "Parque Nacional":
+                experienceParqueNacional.reservation();
+
+                inputExpAdults.addEventListener("change", (_event) => {
+                    experienceParqueNacional.fullPrice();
+                });
+                
+                inputExpChildren.addEventListener("change", (_event) => {
+                    experienceParqueNacional.fullPrice();
+                });
+            break;
+
+            case "Canal Beagle":
+                experienceCanalBeagle.reservation();
+
+                inputExpAdults.addEventListener("change", (_event) => {
+                    experienceCanalBeagle.fullPrice();
+                });
+                
+                inputExpChildren.addEventListener("change", (_event) => {
+                    experienceCanalBeagle.fullPrice();
+                });
+            break;
+        }
+    });
+
     const experienceBookingInformation = getExpInfo();
 
+    //EVENTO PARA ENVIAR INFORMACIÓN DE RESERVA
     experienceBoookingForm.addEventListener("submit", (Event) => {
 
         //STOP EVENT
         Event.preventDefault();
-        
+    
         //SE OBTIENEN DATOS DEL FORMULARIO - GETTING INPUT DATA
-        const experience = selectExperience.value;
+        const experience = selectExperience.value;        
         const guestName = inputExpGuestName.value;
         const guestEmail = inputExpGuestEmail.value;
         const guestTelephone = inputExpGuestTelephone.value;
@@ -127,34 +232,48 @@ experienceBtn.addEventListener("click", (_Event) => {
         const children = inputExpChildren.value;
         const expDate = inputExpDate.value;
     
+        if (experience == "" || guestName == "" || guestTelephone == "" || guestEmail == "" || adults == "" || expDate == ""){
+
+            //MENSAJE DE FALTA DE DATOS
+            Swal.fire('Por favor, complete todos los campos.')
+
+        }else{
+
+            //SE HACE PUSH DE LOS DATOS DE LA RESERVA AL ARRAY - PUSH INPUT DATA INTO AN ARRAY
+            experienceBookingInformation.push({
+                experiencia: experience,
+                nombre: guestName,
+                email: guestEmail,
+                telefono: guestTelephone,
+                cantidadDeAdultos: adults,
+                cantidadDeMenores: children,
+                fecha: expDate,
+            });
+
+            //SE CARGA EL ARRAY AL LOCALSTORAGE - ARRAY IS UPLOADED TO LOCALSTORAGE
+            localStorage.setItem("experienceBookingInformation", JSON.stringify(experienceBookingInformation));
     
-        //SE HACE PUSH DE LOS DATOS DE LA RESERVA AL ARRAY - PUSH INPUT DATA INTO AN ARRAY
-        experienceBookingInformation.push({
-            experiencia: experience,
-            nombre: guestName,
-            email: guestEmail,
-            telefono: guestTelephone,
-            cantidadDeAdultos: adults,
-            cantidadDeMenores: children,
-            fecha: expDate,
-        });
+            //MENSAJE RESERVA COMPLETA
+            Swal.fire({
+                position: 'center',
+                title: 'Su reserva se ha realizado.',
+                showConfirmButton: true,
+            });
 
-        //SE CARGA EL ARRAY AL LOCALSTORAGE - ARRAY IS UPLOADED TO LOCALSTORAGE
-        localStorage.setItem("experienceBookingInformation", JSON.stringify(experienceBookingInformation));
-        
-        alert("Su reserva ha sido realizada.")
+            //SE LIMPIA EL COMUNICADOR DE PRECIOS
+            let container = document.getElementById("container-precios");
+            container.innerHTML = "";
 
-        //SE LIMPIAN LOS INPUTS
-        selectExperience.value = "";
-        inputExpGuestName.value = "";
-        inputExpGuestEmail.value = "";
-        inputExpGuestTelephone.value = "";
-        inputExpAdults.value = "";
-        inputExpChildren.value = "";
-        inputExpDate.value = "";
+            //SE LIMPIAN LOS INPUTS
+            selectExperience.value = "";
+            inputExpGuestName.value = "";
+            inputExpGuestEmail.value = "";
+            inputExpGuestTelephone.value = "";
+            inputExpAdults.value = "";
+            inputExpChildren.value = "";
+            inputExpDate.value = "";
 
-        console.log(experienceBookingInformation);
+            console.log(experienceBookingInformation);
+        };
+    });
 });
-
-});
-
